+++ jest.setup.js
@@ -1,0 +1,7 @@
+// beforeEach(() => {
+//     window.matchMedia = window.matchMedia || (() => ({
+//         matches: false,
+//         addListener: () => {},
+//         removeListener: () => {}
+//     }));
+// });